staging/pkg/mod/golang.org/x/net@v0.0.0-20201110031124-69a78807bb2b/html/testdata/webkit/tests_innerHTML_1.dat

#data
<body><span>
#errors
(1,6): unexpected-start-tag
(1,12): expected-closing-tag-but-got-eof
#document-fragment
body
#document
| <span>

#data
<span><body>
#errors
(1,12): unexpected-start-tag
(1,12): expected-closing-tag-but-got-eof
#document-fragment
body
#document
| <span>

#data
<span><body>
#errors
(1,12): unexpected-start-tag
(1,12): expected-closing-tag-but-got-eof
#document-fragment
div
#document
| <span>

#data
<body><span>
#errors
(1,12): expected-closing-tag-but-got-eof
#document-fragment
html
#document
| <head>
| <body>
|   <span>

#data
<frameset><span>
#errors
(1,10): unexpected-start-tag
(1,16): expected-closing-tag-but-got-eof
#document-fragment
body
#document
| <span>

#data
<span><frameset>
#errors
(1,16): unexpected-start-tag
(1,16): expected-closing-tag-but-got-eof
#document-fragment
body
#document
| <span>

#data
<span><frameset>
#errors
(1,16): unexpected-start-tag
(1,16): expected-closing-tag-but-got-eof
#document-fragment
div
#document
| <span>

#data
<frameset><span>
#errors
(1,16): unexpected-start-tag-in-frameset
(1,16): eof-in-frameset
#document-fragment
html
#document
| <head>
| <frameset>

#data
<table><tr>
#errors
(1,7): unexpected-start-tag
#document-fragment
table
#document
| <tbody>
|   <tr>

#data
</table><tr>
#errors
(1,8): unexpected-end-tag
#document-fragment
table
#document
| <tbody>
|   <tr>

#data
<a>
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
(1,3): eof-in-table
#document-fragment
table
#document
| <a>

#data
<a>
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
(1,3): eof-in-table
#document-fragment
table
#document
| <a>

#data
<a><caption>a
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
(1,13): expected-closing-tag-but-got-eof
#document-fragment
table
#document
| <a>
| <caption>
|   "a"

#data
<a><colgroup><col>
#errors
(1,3): foster-parenting-start-token
(1,18): expected-closing-tag-but-got-eof
#document-fragment
table
#document
| <a>
| <colgroup>
|   <col>

#data
<a><tbody><tr>
#errors
(1,3): foster-parenting-start-tag
#document-fragment
table
#document
| <a>
| <tbody>
|   <tr>

#data
<a><tfoot><tr>
#errors
(1,3): foster-parenting-start-tag
#document-fragment
table
#document
| <a>
| <tfoot>
|   <tr>

#data
<a><thead><tr>
#errors
(1,3): foster-parenting-start-tag
#document-fragment
table
#document
| <a>
| <thead>
|   <tr>

#data
<a><tr>
#errors
(1,3): foster-parenting-start-tag
#document-fragment
table
#document
| <a>
| <tbody>
|   <tr>

#data
<a><th>
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
(1,7): unexpected-cell-in-table-body
#document-fragment
table
#document
| <a>
| <tbody>
|   <tr>
|     <th>

#data
<a><td>
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
(1,7): unexpected-cell-in-table-body
#document-fragment
table
#document
| <a>
| <tbody>
|   <tr>
|     <td>

#data
<table></table><tbody>
#errors
(1,22): unexpected-start-tag
#document-fragment
caption
#document
| <table>

#data
</table><span>
#errors
(1,8): unexpected-end-tag
(1,14): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>

#data
<span></table>
#errors
(1,14): unexpected-end-tag
(1,14): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>

#data
</caption><span>
#errors
(1,10): XXX-undefined-error
(1,16): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>

#data
<span></caption><span>
#errors
(1,16): XXX-undefined-error
(1,22): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><caption><span>
#errors
(1,15): unexpected-start-tag
(1,21): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><col><span>
#errors
(1,11): unexpected-start-tag
(1,17): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><colgroup><span>
#errors
(1,16): unexpected-start-tag
(1,22): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><html><span>
#errors
(1,12): non-html-root
(1,18): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tbody><span>
#errors
(1,13): unexpected-start-tag
(1,19): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><td><span>
#errors
(1,10): unexpected-start-tag
(1,16): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tfoot><span>
#errors
(1,13): unexpected-start-tag
(1,19): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><thead><span>
#errors
(1,13): unexpected-start-tag
(1,19): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><th><span>
#errors
(1,10): unexpected-start-tag
(1,16): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tr><span>
#errors
(1,10): unexpected-start-tag
(1,16): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span></table><span>
#errors
(1,14): unexpected-end-tag
(1,20): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <span>
|   <span>

#data
</colgroup><col>
#errors
(1,11): XXX-undefined-error
#document-fragment
colgroup
#document
| <col>

#data
<a><col>
#errors
(1,3): XXX-undefined-error
#document-fragment
colgroup
#document
| <col>

#data
<caption><a>
#errors
(1,9): XXX-undefined-error
(1,12): unexpected-start-tag-implies-table-voodoo
(1,12): eof-in-table
#document-fragment
tbody
#document
| <a>

#data
<col><a>
#errors
(1,5): XXX-undefined-error
(1,8): unexpected-start-tag-implies-table-voodoo
(1,8): eof-in-table
#document-fragment
tbody
#document
| <a>

#data
<colgroup><a>
#errors
(1,10): XXX-undefined-error
(1,13): unexpected-start-tag-implies-table-voodoo
(1,13): eof-in-table
#document-fragment
tbody
#document
| <a>

#data
<tbody><a>
#errors
(1,7): XXX-undefined-error
(1,10): unexpected-start-tag-implies-table-voodoo
(1,10): eof-in-table
#document-fragment
tbody
#document
| <a>

#data
<tfoot><a>
#errors
(1,7): XXX-undefined-error
(1,10): unexpected-start-tag-implies-table-voodoo
(1,10): eof-in-table
#document-fragment
tbody
#document
| <a>

#data
<thead><a>
#errors
(1,7): XXX-undefined-error
(1,10): unexpected-start-tag-implies-table-voodoo
(1,10): eof-in-table
#document-fragment
tbody
#document
| <a>

#data
</table><a>
#errors
(1,8): XXX-undefined-error
(1,11): unexpected-start-tag-implies-table-voodoo
(1,11): eof-in-table
#document-fragment
tbody
#document
| <a>

#data
<a><tr>
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
#document-fragment
tbody
#document
| <a>
| <tr>

#data
<a><td>
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
(1,7): unexpected-cell-in-table-body
#document-fragment
tbody
#document
| <a>
| <tr>
|   <td>

#data
<a><td>
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
(1,7): unexpected-cell-in-table-body
#document-fragment
tbody
#document
| <a>
| <tr>
|   <td>

#data
<a><td>
#errors
(1,3): unexpected-start-tag-implies-table-voodoo
(1,7): unexpected-cell-in-table-body
#document-fragment
tbody
#document
| <a>
| <tr>
|   <td>

#data
<td><table><tbody><a><tr>
#errors
(1,4): unexpected-cell-in-table-body
(1,21): unexpected-start-tag-implies-table-voodoo
(1,25): eof-in-table
#document-fragment
tbody
#document
| <tr>
|   <td>
|     <a>
|     <table>
|       <tbody>
|         <tr>

#data
</tr><td>
#errors
(1,5): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
<td><table><a><tr></tr><tr>
#errors
(1,14): unexpected-start-tag-implies-table-voodoo
(1,27): eof-in-table
#document-fragment
tr
#document
| <td>
|   <a>
|   <table>
|     <tbody>
|       <tr>
|       <tr>

#data
<caption><td>
#errors
(1,9): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
<col><td>
#errors
(1,5): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
<colgroup><td>
#errors
(1,10): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
<tbody><td>
#errors
(1,7): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
<tfoot><td>
#errors
(1,7): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
<thead><td>
#errors
(1,7): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
<tr><td>
#errors
(1,4): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
</table><td>
#errors
(1,8): XXX-undefined-error
#document-fragment
tr
#document
| <td>

#data
<td><table></table><td>
#errors
#document-fragment
tr
#document
| <td>
|   <table>
| <td>

#data
<td><table></table><td>
#errors
#document-fragment
tr
#document
| <td>
|   <table>
| <td>

#data
<caption><a>
#errors
(1,9): XXX-undefined-error
(1,12): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
<col><a>
#errors
(1,5): XXX-undefined-error
(1,8): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
<colgroup><a>
#errors
(1,10): XXX-undefined-error
(1,13): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
<tbody><a>
#errors
(1,7): XXX-undefined-error
(1,10): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
<tfoot><a>
#errors
(1,7): XXX-undefined-error
(1,10): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
<th><a>
#errors
(1,4): XXX-undefined-error
(1,7): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
<thead><a>
#errors
(1,7): XXX-undefined-error
(1,10): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
<tr><a>
#errors
(1,4): XXX-undefined-error
(1,7): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
</table><a>
#errors
(1,8): XXX-undefined-error
(1,11): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
</tbody><a>
#errors
(1,8): XXX-undefined-error
(1,11): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
</td><a>
#errors
(1,5): unexpected-end-tag
(1,8): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
</tfoot><a>
#errors
(1,8): XXX-undefined-error
(1,11): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
</thead><a>
#errors
(1,8): XXX-undefined-error
(1,11): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
</th><a>
#errors
(1,5): unexpected-end-tag
(1,8): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
</tr><a>
#errors
(1,5): XXX-undefined-error
(1,8): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <a>

#data
<table><td><td>
#errors
(1,11): unexpected-cell-in-table-body
(1,15): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <table>
|   <tbody>
|     <tr>
|       <td>
|       <td>

#data
</select><option>
#errors
(1,9): XXX-undefined-error
#document-fragment
select
#document
| <option>

#data
<input><option>
#errors
(1,7): unexpected-input-in-select
#document-fragment
select
#document
| <option>

#data
<keygen><option>
#errors
(1,8): unexpected-input-in-select
#document-fragment
select
#document
| <option>

#data
<textarea><option>
#errors
(1,10): unexpected-input-in-select
#document-fragment
select
#document
| <option>

#data
</html><!--abc-->
#errors
(1,7): unexpected-end-tag-after-body-innerhtml
#document-fragment
html
#document
| <head>
| <body>
| <!-- abc -->

#data
</frameset><frame>
#errors
(1,11): unexpected-frameset-in-frameset-innerhtml
#document-fragment
frameset
#document
| <frame>

#data
#errors
#document-fragment
html
#document
| <head>
| <body>
